+++ spec/queries/find_ephemera_term_by_label_spec.rb
@@ -4,7 +4,7 @@ require "rails_helper"
 RSpec.describe FindEphemeraTermByLabel do
   subject(:query) { described_class.new(query_service: query_service) }
   let(:term) { FactoryBot.create_for_repository(:ephemera_term, label: "Test", member_of_vocabulary_id: vocab.id) }
-  let(:term2) { FactoryBot.create_for_repository(:ephemera_term, label: "Test", member_of_vocabulary_id: vocab2.id) }
+  let(:term2) { FactoryBot.create_for_repository(:ephemera_term, label: "Test", code: "t", member_of_vocabulary_id: vocab2.id) }
   let(:vocab) { FactoryBot.create_for_repository(:ephemera_vocabulary) }
   let(:vocab2) { FactoryBot.create_for_repository(:ephemera_vocabulary, label: "Test2") }
   let(:query_service) { Valkyrie.config.metadata_adapter.query_service }
@@ -18,6 +18,14 @@ RSpec.describe FindEphemeraTermByLabel do
       term2
       output = query.find_ephemera_term_by_label(label: term.label, parent_vocab_label: vocab.label)
       expect(output.id).to eq term.id
+    end
+    it "can find by code" do
+      term2
+      output = query.find_ephemera_term_by_label(code: "t", parent_vocab_label: vocab2.label)
+      expect(output.id).to eq term2.id
+    end
+    it "errors if neither label nor code is specified" do
+      expect { query.find_ephemera_term_by_label }.to raise_error(ArgumentError)
     end
   end
 end